Use custom ticks in 1-dimension charts also
