+++ components/aggregation/mat/computations.js
@@ -1,0 +1,9 @@
+export function getDatesBetween(startDate, endDate) {
+  const dateArray = new Set()
+  var currentDate = startDate
+  while (currentDate < endDate) {
+    dateArray.add(currentDate.toISOString().slice(0, 10))
+    currentDate.setDate(currentDate.getDate() + 1)
+  }
+  return dateArray
+}
